+++ test/k8s/resource_test.exs
@@ -1,4 +1,6 @@
 defmodule K8s.ResourceTest do
   use ExUnit.Case, async: true
   doctest K8s.Resource
+  doctest K8s.Resource.FieldAccessors
+  doctest K8s.Resource.Utilization
 end